+++ src/actions/index.ts
@@ -1,6 +1,6 @@
 export * from "./audio.js";
 export * from "./images.js";
 export * from "./movie.js";
-export * from "./pdf_puppeteer.js";
+export * from "./pdf.js";
 export * from "./translate.js";
 export * from "./captions.js";
